+++ src/constants/proposals.js
@@ -39,6 +39,29 @@ const TYPES = {
   TEXT: 'TextProposal',
 };
 
+const SUPPORTED_TYPES = [
+  {
+    name: TYPES.CANCEL_SOFTWARE_UPGRADE,
+    text: '/cosmos.upgrade.v1beta1.CancelSoftwareUpgradeProposal',
+  },
+  {
+    name: TYPES.COMMUNITY_POOL_SPEND,
+    text: '/cosmos.distribution.v1beta1.CommunityPoolSpendProposal',
+  },
+  {
+    name: TYPES.PARAMETER_CHANGE,
+    text: '/cosmos.params.v1beta1.ParameterChangeProposal',
+  },
+  {
+    name: TYPES.SOFTWARE_UPGRADE,
+    text: '/cosmos.upgrade.v1beta1.SoftwareUpgradeProposal',
+  },
+  {
+    name: TYPES.TEXT,
+    text: '/cosmos.gov.v1beta1.TextProposal',
+  },
+];
+
 const ICONS = {
   [TYPES.CANCEL_SOFTWARE_UPGRADE]: mdiCogOffOutline,
   [TYPES.COMMUNITY_POOL_SPEND]: mdiCurrencyUsd,
@@ -47,10 +70,20 @@ const ICONS = {
   [TYPES.TEXT]: mdiTextBoxOutline,
 };
 
+const LABELS = {
+  [TYPES.CANCEL_SOFTWARE_UPGRADE]: 'Cancel Software Upgrade Proposal',
+  [TYPES.COMMUNITY_POOL_SPEND]: 'Community Pool Spend Proposal',
+  [TYPES.PARAMETER_CHANGE]: 'Parameter Change Proposal',
+  [TYPES.SOFTWARE_UPGRADE]: 'Software Upgrade Proposal',
+  [TYPES.TEXT]: 'Text Proposal',
+};
+
 export const PROPOSALS = {
   COLORS,
   ICONS,
+  LABELS,
   STATUS,
+  SUPPORTED_TYPES,
   TEXTS,
   TYPES,
 };