+++ src/views/proposals/detail/ProposalsViewDetailDataHeader.vue
@@ -3,17 +3,23 @@
     <v-list-item-content>
       <v-list-item-title class="pb-1 font-weight-bold">
         <v-row>
-          <v-col cols="12" md="9" xl="10">
+          <v-col cols="12" md="8" xl="10">
             <div class="text-wrap text-h6 font-weight-bold">
               <span>#</span>
               <span v-text="id" />
               <span>&nbsp;-&nbsp;</span>
               <span v-text="title" />
             </div>
-            <div class="text-subtitle-1 font-italic" v-text="type" />
+            <div
+              class="text-subtitle-1 font-italic"
+              v-text="PROPOSALS.LABELS[type]"
+            />
           </v-col>
-          <v-col cols="12" md="3" xl="2" class="d-flex justify-md-end">
-            <v-chip class="white--text font-weight-bold" :color="statusColor">
+          <v-col cols="12" md="4" xl="2" class="d-flex justify-md-end">
+            <v-chip
+              class="text-caption white--text font-weight-bold"
+              :color="statusColor"
+            >
               {{ statusText }}
             </v-chip>
           </v-col>
@@ -30,6 +36,9 @@ import { mapGetters } from 'vuex';
 
 export default {
   name: 'ProposalsViewDetailDataHeader',
+  data() {
+    return { PROPOSALS };
+  },
   computed: {
     ...mapGetters('proposals', ['proposal']),
     id() {