+++ src/views/validators/helpers/__tests__/validatorAttendanceCalculator.test.js
@@ -3,7 +3,7 @@ import validatorAttendanceCalculator from '../validatorAttendanceCalculator';
 import { VALIDATORS } from '@/constants';
 
 describe('views/validators/helpers/validatorAttendanceCalculator', () => {
-  test('if "validatorAttendanceCalculator" methods return the defined blocks, the attendance count and percentage', () => {
+  test('if "validatorAttendanceCalculator" methods return the defined blocks and the attendance count', () => {
     const limit = 100;
     const blocks = mockBlocks(limit);
     const validator = mockValidator();
@@ -18,14 +18,6 @@ describe('views/validators/helpers/validatorAttendanceCalculator', () => {
     const attendanceCount =
       validatorAttendanceCalculator.getAttendanceCount(definedBlocks);
     expect(attendanceCount).toBe(limit);
-
-    const attendancePercentage =
-      validatorAttendanceCalculator.getAttendancePercentage(
-        attendanceCount,
-        limit
-      );
-    const splitPercentage = attendancePercentage.split(/[,.]/);
-    expect(splitPercentage[0]).toBe(`${limit}`);
   });
 
   test('if "validatorAttendanceCalculator.getDefinedBlocks" method return an empty list if a validator is not found in validatorsets', () => {