+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.5] - 2022-05-06
+
+### Added
+
+- Add support for many types of proposals
+
+### Changed
+
+- Update proposals chart
+
 ## [3.1.4] - 2022-04-20
 
 ### Fixed
@@ -361,6 +371,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add RFC cross chain navigation
 - Add a switch to selcet a dark theme from navigation drawer
 
+[3.1.5]: https://github.com/commercionetwork/almerico/compare/v3.1.4...v3.1.5
+[3.1.4]: https://github.com/commercionetwork/almerico/compare/v3.1.3...v3.1.4
 [3.1.3]: https://github.com/commercionetwork/almerico/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/commercionetwork/almerico/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/commercionetwork/almerico/compare/v3.1.0...v3.1.1